apps/backend/pyproject.toml
[project]
name = "voicerag-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
    "livekit-api>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "openai>=1.50.0",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.2",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
